fix(mink-core.bb): set current configure options
Set the list of current configurable options to be visible and thus,
configurable directly from the recipe. Also, mind the --with-boost
handling of boost paths in order to avoid QA complaints and search for
includes/libraries on the host machine, rather then build system.

Remove uneccessary variables for libraries as well.

Signed-off-by: Davor Popovic <davor.popeye@gmail.com>